wip: update tests, @nuxtjs/svg fails tests.
